Move agora-results logic to agora_results module
Moves all the logic to the agora_results module to make it easier to use
this as a module and not just as a script.

This simplify a bit all the tests because now it's not needed to call
the subprocess and we can mock better the parameters.

The agora-results script now just parses the input and calls the main
method, but everything else is inside the agora_results.main module.